% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{flock_detailed}
\alias{flock_detailed}
\title{executes flocking and returns a dataframe with the location data}
\usage{
flock_detailed(
  no_iter,
  init_boids,
  save_locations_path,
  sample_rate,
  init_width,
  init_height,
  sensory_distance,
  allignment_coef,
  cohesion_coef,
  separation_coef,
  allignment_trs_coef,
  cohesion_trs_coef,
  separation_trs_coef,
  min_speed,
  max_speed,
  max_steering
)
}
\description{
executes flocking and returns a dataframe with the location data
}
