% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{flock_detailed_no_return}
\alias{flock_detailed_no_return}
\title{executes flocking and returns a dataframe with the location data}
\usage{
flock_detailed_no_return(
  no_iter,
  init_boids,
  save_locations_path,
  sample_rate,
  init_width,
  init_height,
  sensory_distance,
  alignment_coef,
  cohesion_coef,
  separation_coef,
  alignment_trs_coef,
  cohesion_trs_coef,
  separation_trs_coef,
  min_speed,
  agent_steering,
  max_speed,
  max_steering,
  dbscan_clustering,
  boundary_config,
  distance_config,
  field_of_vision,
  rules_impl,
  wander_on,
  wander_coef,
  wander_rate,
  wander_radius,
  wander_distance,
  baseline_speed,
  wander_random
)
}
\description{
executes flocking and returns a dataframe with the location data
}
