% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{flock_return}
\alias{flock_return}
\title{executes flocking and returns a dataframe with the location data}
\usage{
flock_return(
  no_iter,
  init_boids,
  save_locations_path,
  sample_rate,
  init_width,
  init_height
)
}
\description{
executes flocking and returns a dataframe with the location data
}
